requirements.txt
requests>=2.25.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3
numpy>=1.19.5
scikit-learn>=1.0.0

